Create README.md
The primary focus of this software design project was to provide a solution to the redundant, time-consuming task of physically purchasing groceries, by developing an online grocery delivery system, also known as an “InstaCart.” Today, the average person lives a relatively busy and exhausting life, and often hectic schedules prevent one from having the time to do tasks as trivial as grocery shopping. Thus, a system that would allow one to have groceries delivered to their doorstep in a minimum period would bring great ease into one’s life. Keeping this in mind, we started this project, which was to generate a software design project that would provide an online shopping platform where individuals will be able to manage their groceries more efficiently. The scope of this project will, however, be limited to different types of Milk, and can, eventually, be expanded to several other types of groceries. The idea of this software design was acquired from the leading online shopping platform like eBay, Amazon, etc. where the chief consideration stays on non-food items and not particularly groceries. This module, however, concentrates only on groceries, making it easier to access than the other commercial websites. Moreover, purchases on this system are carried out locally, and subsequently, take the little time to be delivered with the minimum shipping charges. Consumers will have the facility to buy products directly from the website using this module. The Manager/Admin is responsible for updating the grocery items prices on a regular basis. For this project’s reference, the manager’s Username and Password are Admin and 123 respectively.

The project was completed in a group of three students